Fix crash on malformed query params and add setup-failure diagnostics
Ran a full sweep of every page in the app (all 114 admin/org-admin/
worker/public files) with real data and confirmed the code itself
works correctly end to end. Two real gaps turned up during that sweep
and while reproducing common first-run mistakes:

1. clean_int($_GET['id'] ?? 0) returns null (not 0) when the query
   param is present but empty (e.g. a stray "?id=" with no value),
   since the ?? fallback only applies when the key is entirely absent.
   That null then hit a non-nullable `int $id` parameter on
   tenant_find()/tenant_update() and threw an uncaught TypeError - a
   500 instead of a graceful "not found" redirect. Wrapped all 26
   call sites across 23 files with (int) so they always resolve to a
   real int.

2. When the database isn't reachable yet (no .env, wrong credentials,
   migrations not run) or vendor/ hasn't been installed (composer
   install not run), the app fell through to a bare "Something went
   wrong" with zero information - and since nearly every page touches
   the database immediately, this made the entire app look broken
   rather than "not finished being set up yet". The global exception
   handler in includes/bootstrap.php now recognizes these two specific
   failure modes and shows plain-language setup instructions instead,
   regardless of APP_DEBUG.

Verified both fixes directly: served an empty ?id= param and confirmed
a clean redirect instead of a crash; took the database offline and
moved vendor/ aside independently and confirmed each now renders a
readable "what to do next" page instead of a blank failure. Full
25-test suite still passes.

+++ admin/subscriptions/plans.php
@@ -3,7 +3,7 @@ require_once __DIR__ . '/../../includes/bootstrap.php';
 
 $admin = require_platform_admin();
 
-$editId = clean_int($_GET['edit'] ?? 0);
+$editId = (int) clean_int($_GET['edit'] ?? 0);
 $editingPlan = $editId ? db_one('SELECT * FROM subscription_plans WHERE id = :id', ['id' => $editId]) : null;
 
 $errors = [];
